fixed icon styling
