+++ src/components/UI/Form/WhatsAppEditor/WhatsAppEditor.tsx
@@ -1,11 +1,9 @@
-import { useState } from 'react';
 import { RichUtils, getDefaultKeyBinding, Modifier, EditorState, Editor } from 'draft-js';
-import { IconButton, ClickAwayListener } from '@mui/material';
 import ReactResizeDetector from 'react-resize-detector';
 import { useTranslation } from 'react-i18next';
 
 import { getPlainTextFromEditor } from 'common/RichEditor';
-import { EmojiPicker } from 'components/UI/EmojiPicker/EmojiPicker';
+
 import styles from './WhatsAppEditor.module.css';
 
 interface WhatsAppEditorProps {
@@ -16,6 +14,27 @@ interface WhatsAppEditorProps {
   readOnly?: boolean;
 }
 
+export const updatedValue = (input: any, editorState: any, isEmoji: boolean = false) => {
+  const editorContentState = editorState.getCurrentContent();
+  const editorSelectionState: any = editorState.getSelection();
+  const ModifiedContent = Modifier.replaceText(
+    editorContentState,
+    editorSelectionState,
+    isEmoji ? input.native : input
+  );
+  let updatedEditorState = EditorState.push(editorState, ModifiedContent, 'insert-characters');
+  if (!isEmoji) {
+    const editorSelectionStateMod = updatedEditorState.getSelection();
+    const updatedSelection = editorSelectionStateMod.merge({
+      anchorOffset: editorSelectionStateMod.getAnchorOffset() - 1,
+      focusOffset: editorSelectionStateMod.getFocusOffset() - 1,
+    });
+    updatedEditorState = EditorState.forceSelection(updatedEditorState, updatedSelection);
+  }
+
+  return updatedEditorState;
+};
+
 export const WhatsAppEditor = ({
   setEditorState,
   sendMessage,
@@ -23,31 +42,10 @@ export const WhatsAppEditor = ({
   handleHeightChange,
   readOnly = false,
 }: WhatsAppEditorProps) => {
-  const [showEmojiPicker, setShowEmojiPicker] = useState(false);
   const { t } = useTranslation();
 
   const handleChange = (editorStateChange: any) => {
     setEditorState(editorStateChange);
-  };
-
-  const updateValue = (input: any, isEmoji: boolean = false) => {
-    const editorContentState = editorState.getCurrentContent();
-    const editorSelectionState: any = editorState.getSelection();
-    const ModifiedContent = Modifier.replaceText(
-      editorContentState,
-      editorSelectionState,
-      isEmoji ? input.native : input
-    );
-    let updatedEditorState = EditorState.push(editorState, ModifiedContent, 'insert-characters');
-    if (!isEmoji) {
-      const editorSelectionStateMod = updatedEditorState.getSelection();
-      const updatedSelection = editorSelectionStateMod.merge({
-        anchorOffset: editorSelectionStateMod.getAnchorOffset() - 1,
-        focusOffset: editorSelectionStateMod.getFocusOffset() - 1,
-      });
-      updatedEditorState = EditorState.forceSelection(updatedEditorState, updatedSelection);
-    }
-    setEditorState(updatedEditorState);
   };
 
   const handleKeyCommand = (command: string, editorStateChange: any) => {
@@ -63,9 +61,9 @@ export const WhatsAppEditor = ({
     }
 
     if (command === 'bold') {
-      updateValue('**');
+      setEditorState(updatedValue('**', editorState));
     } else if (command === 'italic') {
-      updateValue('__');
+      setEditorState(updatedValue('__', editorState));
     } else {
       const newState = RichUtils.handleKeyCommand(editorStateChange, command);
       if (newState) {
@@ -84,64 +82,24 @@ export const WhatsAppEditor = ({
     return getDefaultKeyBinding(e);
   };
 
-  const handleClickAway = () => {
-    setShowEmojiPicker(false);
-  };
-
-  const emojiStyles: any = {
-    position: 'absolute',
-    bottom: '60px',
-    right: '-150px',
-    zIndex: 100,
-  };
-
-  if (window.innerWidth <= 768) {
-    emojiStyles.right = '5%';
-  }
-
   return (
-    <>
-      <ReactResizeDetector
-        data-testid="resizer"
-        handleHeight
-        onResize={(width: any, height: any) => handleHeightChange(height - 40)} // 40 is the initial height
-      >
-        <div className={styles.Editor}>
-          <Editor
-            data-testid="editor"
-            editorState={editorState}
-            onChange={handleChange}
-            handleKeyCommand={handleKeyCommand}
-            keyBindingFn={keyBindingFn}
-            placeholder={t('Type a message...')}
-            readOnly={readOnly}
-          />
-        </div>
-      </ReactResizeDetector>
-      <ClickAwayListener onClickAway={handleClickAway}>
-        <div>
-          <div className={styles.EmojiButton}>
-            <IconButton
-              data-testid="emoji-picker"
-              color="primary"
-              aria-label="pick emoji"
-              component="span"
-              onClick={() => setShowEmojiPicker(!showEmojiPicker)}
-            >
-              <span role="img" aria-label="pick emoji">
-                😀
-              </span>
-            </IconButton>
-          </div>
-          {showEmojiPicker ? (
-            <EmojiPicker
-              onEmojiSelect={(emoji: any) => updateValue(emoji, true)}
-              displayStyle={emojiStyles}
-            />
-          ) : null}
-        </div>
-      </ClickAwayListener>
-    </>
+    <ReactResizeDetector
+      data-testid="resizer"
+      handleHeight
+      onResize={(width: any, height: any) => handleHeightChange(height - 40)} // 40 is the initial height
+    >
+      <div className={styles.Editor}>
+        <Editor
+          data-testid="editor"
+          editorState={editorState}
+          onChange={handleChange}
+          handleKeyCommand={handleKeyCommand}
+          keyBindingFn={keyBindingFn}
+          placeholder={t('Type a message...')}
+          readOnly={readOnly}
+        />
+      </div>
+    </ReactResizeDetector>
   );
 };
