+++ src/containers/Chat/ChatMessages/ChatInput/ChatInput.module.css
@@ -18,8 +18,6 @@
   justify-content: space-between;
   height: fit-content;
   max-height: 200px;
-  overflow-y: scroll;
-  padding-right: 5px;
 }
 
 .Rounded {
@@ -93,19 +91,14 @@
 }
 
 .SendButtonContainer {
-  position: absolute;
-  bottom: 10px;
-  right: 35px;
+  padding: 0 10px;
 }
 
 .EmojiButton {
-  position: absolute;
-  bottom: 14px;
-  right: 78px;
+  padding: 0px 10px !important;
 }
 
 .Editor {
-  width: 75%;
   min-height: 40px;
 }
 
@@ -202,33 +195,9 @@
   padding-top: 20px;
 }
 
-.AttachmentIcon {
-  position: absolute !important;
-  bottom: 7px;
-  right: 122px;
-}
-
 .AttachmentIcon svg {
   height: 24px;
   width: 24px;
-}
-
-.VariableIcon {
-  position: absolute !important;
-  bottom: 4px;
-  right: 165px;
-}
-
-.CrossIcon {
-  position: absolute !important;
-  bottom: 4px;
-  right: 210px;
-}
-
-.CrossIconWithVariable {
-  position: absolute !important;
-  bottom: 4px;
-  right: 165px;
 }
 
 .CrossIcon svg,
@@ -241,39 +210,14 @@ svg.SendIcon {
   margin-top: 2px;
 }
 
+.Icons {
+  display: flex;
+  align-items: center;
+}
+
 @media only screen and (max-width: 768px) {
   .QuickSend {
     font-size: 12px;
-  }
-
-  .CrossIcon {
-    position: absolute !important;
-    bottom: -5%;
-    right: 60%;
-  }
-
-  .CrossIconWithVariable {
-    position: absolute !important;
-    bottom: -5%;
-    right: 60%;
-  }
-
-  .VariableIcon {
-    position: absolute !important;
-    bottom: -5%;
-    right: 50%;
-  }
-
-  .AttachmentIcon {
-    position: absolute !important;
-    bottom: 0px;
-    right: 30%;
-  }
-
-  .SendButtonContainer {
-    position: absolute;
-    bottom: 5px;
-    right: 10%;
   }
 
   .ChatInputElements {
@@ -285,7 +229,5 @@ svg.SendIcon {
     justify-content: space-between;
     height: fit-content;
     max-height: 200px;
-    overflow-y: scroll;
-    padding-right: 5px;
   }
 }
